+++ R/map.R
@@ -32,9 +32,16 @@
 #' @section Results:
 #'
 #'     To collect the results of the map operation, use \code{x[]} on a
-#'     mirai_map \sQuote{x}. This will wait for all asynchronous
+#'     mirai_map \code{x}. This will wait for all asynchronous
 #'     operations to complete if still in progress (blocking, although
 #'     user-interruptible).
+#'
+#'     To collect the results whilst showing a text progress indicator, use
+#'     \code{x[.progress]}.
+#'
+#'     To collect the results with early stopping, which stops at the first
+#'     failure and aborts all remaining in-progress operations, use
+#'     \code{x[.stop]}.
 #'
 #' @details Sends each application of function \code{.f} on an element of
 #'     \code{.x} for computation in a separate \code{\link{mirai}} call.
@@ -61,12 +68,15 @@
 #' # progress indicator counts up to 4 seconds
 #' with(
 #'   daemons(4, dispatcher = FALSE),
-#'   mirai_map(1:4, Sys.sleep)[]
+#'   mirai_map(1:4, Sys.sleep)[.progress]
 #' )
 #'
 #' # creates 3 ephemeral daemons as daemons not set
-#' # second element returns a 'miraiError'
-#' mirai_map(list(a = 1, b = "a", c = 3), sum)[]
+#' # stops early when second element returns an error
+#' tryCatch(
+#'   mirai_map(list(a = 1, b = "a", c = 3), sum)[.stop],
+#'   error = identity
+#' )
 #'
 #' ml <- mirai_map(c(a = 2, b = 3, c = 4), rnorm)
 #' ml
@@ -89,47 +99,20 @@ mirai_map <- function(.x, .f, ..., .args = list(), .compute = "default") {
 
 }
 
-#' mirai Collect
+#' @export
 #'
-#' Collects the results from \code{mirai_map} or any list of \sQuote{mirai}
-#'     objects, waiting for resolution if still in progress.
-#'
-#' @param x a list of \sQuote{mirai} objects.
-#' @param progress [default FALSE] if TRUE, reports progress via a simple text
-#'     progress indicator.
-#' @param stop [default FALSE] errors are returned as \sQuote{miraiError} /
-#'     \sQuote{errorValue} as the case may be, allowing recovery from partial
-#'     failure. If TRUE, performs early stopping as soon as an error is
-#'     encountered, with remaining in-progress computations aborted.
-#'
-#' @return A list, the same length as \sQuote{x}, preserving names.
-#'
-#' @details This function will wait for all asynchronous operations to
-#'     complete if still in progress (blocking, although user-interruptible).
-#'
-#'     Optionally shows a simple text progress indicator.
-#'
-#'     Allows for early stopping, which stops at the first failure and aborts
-#'     all remaining in-progress computations.
-#'
-#' @keywords internal
-#'
-collect <- function(x, progress = FALSE, stop = FALSE) {
+`[.mirai_map` <- function(x, i) {
 
-  progress || stop || return(aio_collect_(x))
+  missing(i) && return(aio_collect_(x))
 
+  .expr <- i
+  xi <- i <- 0L
   xlen <- length(x)
+  eval(.expr)
   for (i in seq_len(xlen)) {
-    if (progress)
-      cat(sprintf("\r[ %d / %d .... ]", i - 1L, xlen), file = stderr())
-    res <- aio_collect_(x[[i]])
-    stop && is_error_value(res) && {
-      lapply(x, stop_aio)
-      stop(res)
-    }
+    xi <- aio_collect_(x[[i]])
+    eval(.expr)
   }
-  if (progress)
-    cat(sprintf("\r[ %d / %d done ]\n", xlen, xlen), file = stderr())
 
   lapply(x, .subset2, "value")
 
@@ -137,13 +120,17 @@ collect <- function(x, progress = FALSE, stop = FALSE) {
 
 #' @export
 #'
-`[.mirai_map` <- function(x, i) aio_collect_(x)
-
-#' @export
-#'
 print.mirai_map <- function(x, ...) {
 
-  cat(sprintf("< mirai map >\n - items: %d\n", length(x)), file = stdout())
+  cat(sprintf("< mirai map | %d items >\n", length(x)), file = stdout())
   invisible(x)
 
 }
+
+#' @export
+#'
+.progress <- quote(cat(if (i < xlen) sprintf("\r[ %d / %d .... ]", i, xlen) else sprintf("\r[ %d / %d done ]\n", i, xlen), file = stderr()))
+
+#' @export
+#'
+.stop <- quote(if (is_error_value(xi)) { lapply(x, stop_mirai); stop(xi, call. = FALSE) })