+++ NEWS.md
@@ -2,8 +2,9 @@
 
 * Adds `x[]` as a more efficient equivalent of `call_mirai_(x)$data`.
 * Adds `mirai_map()` for asynchronous parallel/distributed map using mirai. Allows recovery from partial failure or else early stopping, together with optional progress reporting.
-  + Accessing `x[]` of a mirai_map `x` waits for and returns the results of the map.
-* Internal efficiency enhancements.
+  + For a mirai_map `x`, the method `x[]`  collects the results of the map, waiting for all asynchronous operations to complete.
+  + `x[.progress]` collects the results whilst showing a text progress bar
+  + `x[.stop]` collects the results applying early-stopping, which stops at the first error, and aborts remaining in-progress operations.
 * Requires `nanonext` >= [1.0.0.9008].
 
 # mirai 1.0.0